+++ src/main/java/raidzero/robot/Robot.java
@@ -7,7 +7,7 @@ package raidzero.robot;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import raidzero.robot.subsystems.lighting.ArmStrip;
+import raidzero.robot.subsystems.LEDStrip.ArmStrip;
 import raidzero.robot.subsystems.telescopingarm.Arm;
 import raidzero.robot.subsystems.telescopingarm.CoralIntake;
